Do step 15 - Remove the Unit Test for the ShoppingCart controller and remove the first version of ShoppingCartRepsitoryTest. Sorry, I forgot to remove it.
