+++ .env.template
@@ -1,0 +1,15 @@
+MYSQLDB_USER=
+MYSQLDB_ROOT_PASSWORD=
+MYSQLDB_DATABASE=
+
+MYSQLDB_LOCAL_PORT=
+MYSQLDB_DOCKER_PORT=
+
+SPRING_LOCAL_PORT=
+SPRING_DOCKER_PORT=
+
+DEBUG_PORT=
+
+JWT_EXPIRATION_TIME=
+JWT_SECRET=
+
